requirements.txt
apify-client==1.8.1
google-auth==2.29.0
google-api-python-client==2.126.0
requests==2.32.3
